After code review
