+++ portal/src/app/components/protocol-query/translate/en.ts
@@ -10,7 +10,6 @@ export const PROTOCOL_QUERY_EN = {
         EXPORT: 'Export',
     },
     FIELDS_LABEL: {
-        PLACEHOLDER: '',
         REQUIRED: '*',
         EXPERIMENT_NAME: 'Experiment name',
         ITERATIONS_NUMBER: 'Number of iterations',
