+++ portal/src/app/shared/components/selector-custom-option/components/index.ts
@@ -1,0 +1,1 @@
+export * from './CustomInput';
